pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "several"
version = "0.1.0"
description = "Universal AI Agent Orchestrator"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "Several Contributors" }
]
dependencies = [
  "textual>=0.52.0",
  "rich>=13.0.0",
  "PyYAML>=6.0",
  "typer>=0.9.0",
  "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-cov>=4.0.0",
  "pytest-asyncio>=0.21.0",
  "black>=23.0.0",
  "ruff>=0.1.0",
  "mypy>=1.0.0",
]

[project.scripts]
several = "several.cli:main"
svr = "several.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
  "-v",
  "--strict-markers",
  "--cov=several",
  "--cov-report=term-missing",
]
markers = [
  "unit: Unit tests",
  "integration: Integration tests",
  "e2e: End-to-end tests",
  "security: Security tests",
  "slow: Slow tests",
]
